Create informatica_2014_level1_pyramid matrix
This code finds the minimum path from (0,0) to (m,n). For example, in the sample input, starts from index inp[0][0](which is 5) and checks the min next val by comparing right next(7), below next(1) and diagonal next(8) values. And moves the current position to the minimum val(1) here. Likewise, moves till the inp[m][n](8) here. And returns the output as the sum of the total path.